pyproject.toml
[tool.poetry]
name = "spacetraders-api"
version = "0.1.0"
description = "https://docs.spacetraders.io/quickstart/new-game"
authors = ["Ashley Felton <ashley@ropable.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "spacetraders_api"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.30.0"
python-dotenv = "^1.0.0"
pydantic = "2.5.1"
ratelimit = "2.2.1"
humanize = "4.9.0"
django = "4.2.8"
dj-database-url = "2.1.0"
argon2-cffi = "^23.1.0"
django-extensions = "3.2.3"
psycopg2 = "2.9.9"

[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
